fastapi-server/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
blinker==1.6.3
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.1.7
climate==0.1.0
dataclasses-json==0.5.14
decorator==5.1.1
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.104.1
frozenlist==1.3.3
fsspec==2023.10.0
greenlet==3.0.0
h11==0.14.0
hurry.filesize==0.9
idna==3.3
ipynbcompress==0.4.0
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.17.3
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
nltk==3.8.1
numpy==1.25.0
openai==1.50.2
packaging==23.2
pandas==2.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.15.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
pytube==15.0.0
pytz==2023.3
PyYAML==6.0.1
regex==2023.10.3
requests==2.27.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
stack-data==0.6.2
starlette==0.27.0
tenacity==8.2.3
tiktoken==0.5.1
tqdm==4.65.0
traitlets==5.9.0
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2023.3
urllib3==1.26.8
uvicorn==0.23.2
wcwidth==0.2.6
Werkzeug==3.0.0
yarl==1.9.2
youtube-transcript-api==0.6.1
